+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Fixed
+- settings: valid ip and network default values were wrongly reported having the
+  wrong type.
+
 ## [1.1.0] - 2024-10-18
 
 ### Added
@@ -72,7 +78,7 @@ and this project adheres to
 
 ## [1.0.0] - 2024-04-08
 
-[unreleased]: https://github.com/rackslab/RFL/compare/v1.0.1...HEAD
+[unreleased]: https://github.com/rackslab/RFL/compare/v1.1.0...HEAD
 [1.1.0]: https://github.com/rackslab/RFL/releases/tag/v1.1.0
 [1.0.3]: https://github.com/rackslab/RFL/releases/tag/v1.0.3
 [1.0.2]: https://github.com/rackslab/RFL/releases/tag/v1.0.2