+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [1.1.1] - 2024-11-05
 
 ### Added
 - web: Add warning log entry in case of JWT decode error.
@@ -81,7 +81,8 @@ and this project adheres to
 
 ## [1.0.0] - 2024-04-08
 
-[unreleased]: https://github.com/rackslab/RFL/compare/v1.1.0...HEAD
+[unreleased]: https://github.com/rackslab/RFL/compare/v1.1.1...HEAD
+[1.1.1]: https://github.com/rackslab/RFL/releases/tag/v1.1.1
 [1.1.0]: https://github.com/rackslab/RFL/releases/tag/v1.1.0
 [1.0.3]: https://github.com/rackslab/RFL/releases/tag/v1.0.3
 [1.0.2]: https://github.com/rackslab/RFL/releases/tag/v1.0.2